move all helper methods to url_pattern.cpp
