+++ src/url_pattern_helpers.cpp
@@ -6,6 +6,198 @@
 
 namespace ada::url_pattern_helpers {
 
+std::tuple<std::string, std::vector<std::string>>
+generate_regular_expression_and_name_list(
+    std::vector<url_pattern_part>& part_list,
+    url_pattern_compile_component_options options) {
+  // Let result be "^"
+  std::string result = "^";
+
+  // Let name list be a new list
+  std::vector<std::string> name_list{};
+  const std::string full_wildcard_regexp_value = ".*";
+
+  // For each part of part list:
+  for (const url_pattern_part& part : part_list) {
+    // If part's type is "fixed-text":
+    if (part.type == url_pattern_part_type::FIXED_TEXT) {
+      // If part's modifier is "none"
+      if (part.modifier == url_pattern_part_modifier::NONE) {
+        // Append the result of running escape a regexp string given part's
+        // value
+        result += escape_regexp_string(part.value);
+      } else {
+        // A "fixed-text" part with a modifier uses a non capturing group
+        // (?:<fixed text>)<modifier>
+        // Append "(?:" to the end of result.
+        result.append("(?:");
+        // Append the result of running escape a regexp string given part’s
+        // value to the end of result.
+        result.append(escape_regexp_string(part.value));
+        // Append ")" to the end of result.
+        result.append(")");
+        // Append the result of running convert a modifier to a string given
+        // part’s modifier to the end of result.
+        result.append(convert_modifier_to_string(part.modifier));
+      }
+      continue;
+    }
+
+    // Assert: part's name is not the empty string
+    ADA_ASSERT_TRUE(!part.name.empty());
+
+    // Append part's name to name list
+    name_list.push_back(part.name);
+
+    // Let regexp value be part's value
+    std::string regexp_value = part.value;
+
+    // If part's type is "segment-wildcard"
+    if (part.type == url_pattern_part_type::SEGMENT_WILDCARD) {
+      // then set regexp value to the result of running generate a segment
+      // wildcard regexp given options.
+      regexp_value = generate_segment_wildcard_regexp(options);
+    }
+    // Otherwise if part's type is "full-wildcard"
+    else if (part.type == url_pattern_part_type::FULL_WILDCARD) {
+      // then set regexp value to full wildcard regexp value.
+      regexp_value = full_wildcard_regexp_value;
+    }
+
+    // If part's prefix is the empty string and part's suffix is the empty
+    // string
+    if (part.prefix.empty() && part.suffix.empty()) {
+      // If part's modifier is "none" or "optional"
+      if (part.modifier == url_pattern_part_modifier::NONE ||
+          part.modifier == url_pattern_part_modifier::OPTIONAL) {
+        // (<regexp value>)<modifier>
+        result += "(" + regexp_value + ")" +
+                  convert_modifier_to_string(part.modifier);
+      } else {
+        // ((?:<regexp value>)<modifier>)
+        result += "((?:" + regexp_value + ")" +
+                  convert_modifier_to_string(part.modifier) + ")";
+      }
+      continue;
+    }
+
+    // If part's modifier is "none" or "optional"
+    if (part.modifier == url_pattern_part_modifier::NONE ||
+        part.modifier == url_pattern_part_modifier::OPTIONAL) {
+      // (?:<prefix>(<regexp value>)<suffix>)<modifier>
+      result += "(?:" + escape_regexp_string(part.prefix) + "(" + regexp_value +
+                ")" + escape_regexp_string(part.suffix) + ")" +
+                convert_modifier_to_string(part.modifier);
+      continue;
+    }
+
+    // Assert: part's modifier is "zero-or-more" or "one-or-more"
+    ADA_ASSERT_TRUE(part.modifier == url_pattern_part_modifier::ZERO_OR_MORE ||
+                    part.modifier == url_pattern_part_modifier::ONE_OR_MORE);
+
+    // Assert: part's prefix is not the empty string or part's suffix is not the
+    // empty string
+    ADA_ASSERT_TRUE(!part.prefix.empty() || !part.suffix.empty());
+
+    // (?:<prefix>((?:<regexp value>)(?:<suffix><prefix>(?:<regexp
+    // value>))*)<suffix>)?
+    // Append "(?:" to the end of result.
+    result.append("(?:");
+    // Append the result of running escape a regexp string given part’s prefix
+    // to the end of result.
+    result.append(escape_regexp_string(part.prefix));
+    // Append "((?:" to the end of result.
+    result.append("((?:");
+    // Append regexp value to the end of result.
+    result.append(regexp_value);
+    // Append ")(?:" to the end of result.
+    result.append(")(?:");
+    // Append the result of running escape a regexp string given part’s suffix
+    // to the end of result.
+    result.append(escape_regexp_string(part.suffix));
+    // Append the result of running escape a regexp string given part’s prefix
+    // to the end of result.
+    result.append(escape_regexp_string(part.prefix));
+    // Append "(?:" to the end of result.
+    result.append("(?:");
+    // Append regexp value to the end of result.
+    result.append(regexp_value);
+    // Append "))*)" to the end of result.
+    result.append("))*)");
+    // Append the result of running escape a regexp string given part’s suffix
+    // to the end of result.
+    result.append(escape_regexp_string(part.suffix));
+    // Append ")" to the end of result.
+    result.append(")");
+
+    // If part's modifier is "zero-or-more" then append "?" to the end of result
+    if (part.modifier == url_pattern_part_modifier::ZERO_OR_MORE) {
+      result += "?";
+    }
+  }
+
+  // Append "$" to the end of result
+  result += "$";
+
+  // Return (result, name list)
+  return {result, name_list};
+}
+
+bool is_ipv6_address(std::string_view input) noexcept {
+  // If input’s code point length is less than 2, then return false.
+  if (input.size() < 2) return false;
+
+  // Let input code points be input interpreted as a list of code points.
+  // If input code points[0] is U+005B ([), then return true.
+  if (input.front() == '[') return true;
+  // If input code points[0] is U+007B ({) and input code points[1] is U+005B
+  // ([), then return true.
+  if (input.starts_with("{[")) return true;
+  // If input code points[0] is U+005C (\) and input code points[1] is U+005B
+  // ([), then return true.
+  return input.starts_with("\\[");
+}
+
+std::string convert_modifier_to_string(url_pattern_part_modifier modifier) {
+  // TODO: Optimize this.
+  switch (modifier) {
+      // If modifier is "zero-or-more", then return "*".
+    case url_pattern_part_modifier::ZERO_OR_MORE:
+      return "*";
+    // If modifier is "optional", then return "?".
+    case url_pattern_part_modifier::OPTIONAL:
+      return "?";
+    // If modifier is "one-or-more", then return "+".
+    case url_pattern_part_modifier::ONE_OR_MORE:
+      return "+";
+    // Return the empty string.
+    default:
+      return "";
+  }
+}
+
+std::string generate_segment_wildcard_regexp(
+    url_pattern_compile_component_options options) {
+  // Let result be "[^".
+  std::string result = "[^";
+  // Append the result of running escape a regexp string given options’s
+  // delimiter code point to the end of result.
+  result.append(escape_regexp_string(options.get_delimiter()));
+  // Append "]+?" to the end of result.
+  result.append("]+?");
+  // Return result.
+  ada_log("generate_segment_wildcard_regexp result: ", result);
+  return result;
+}
+
+bool protocol_component_matches_special_scheme(
+    url_pattern_component& component) {
+  auto regex = component.regexp;
+  return std::regex_match("http", regex) || std::regex_match("https", regex) ||
+         std::regex_match("ws", regex) || std::regex_match("wss", regex) ||
+         std::regex_match("ftp", regex);
+}
+
 inline std::optional<url_pattern_errors>
 constructor_string_parser::compute_protocol_matches_special_scheme_flag() {
   ada_log(
